Command split: Implement the "convert" command in a separate class
